+++ Biobase/Secondary.hs
@@ -1,19 +1,19 @@
 
 module Biobase.Secondary
-  ( module Biobase.Secondary.Basepair
-  , module Biobase.Secondary.Constraint
+--  ( module Biobase.Secondary.Basepair
+  ( module Biobase.Secondary.Constraint
   , module Biobase.Secondary.Diagrams
   , module Biobase.Secondary.Isostericity
   , module Biobase.Secondary.Pseudoknots
   , module Biobase.Secondary.Structure
-  , module Biobase.Secondary.Vienna
+--  , module Biobase.Secondary.Vienna
   ) where
 
-import Biobase.Secondary.Basepair
+--import Biobase.Secondary.Basepair
 import Biobase.Secondary.Constraint
 import Biobase.Secondary.Diagrams
 import Biobase.Secondary.Isostericity
 import Biobase.Secondary.Pseudoknots
 import Biobase.Secondary.Structure
-import Biobase.Secondary.Vienna
+--import Biobase.Secondary.Vienna